+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/property/Type.kt
@@ -28,8 +28,6 @@ import io.verik.compiler.core.common.CardinalDeclaration
 import io.verik.compiler.core.common.CardinalUnresolvedDeclaration
 import io.verik.compiler.core.common.Core
 import io.verik.compiler.core.common.CoreClassDeclaration
-import io.verik.compiler.core.common.OptionalDeclaration
-import io.verik.compiler.core.common.OptionalUnresolvedDeclaration
 import io.verik.compiler.message.Messages
 import io.verik.compiler.message.SourceLocation
 import io.verik.compiler.target.common.TargetClassDeclaration
@@ -69,25 +67,10 @@ class Type(
         }
     }
 
-    fun isOptionalType(): Boolean {
-        return when (val reference = reference) {
-            is ETypeParameter -> reference.type.isOptionalType()
-            is ETypeAlias -> reference.type.isOptionalType()
-            is OptionalDeclaration -> true
-            else -> false
-        }
-    }
-
     fun isResolved(): Boolean {
-        if (reference is CardinalUnresolvedDeclaration || reference is OptionalUnresolvedDeclaration)
+        if (reference is CardinalUnresolvedDeclaration)
             return false
         return arguments.all { it.isResolved() }
-    }
-
-    fun isSpecialized(): Boolean {
-        if (isCardinalType() && reference !is CardinalConstantDeclaration)
-            return false
-        return arguments.all { it.isSpecialized() }
     }
 
     fun asCardinalValue(element: EElement): Int {