+++ verik-compiler/src/main/kotlin/io/verik/compiler/specialize/ExpressionSpecializer.kt
@@ -143,6 +143,7 @@ object ExpressionSpecializer {
             val callExpressionTypeParameterContext = TypeParameterContext.getFromTypeArguments(
                 callExpression.typeArguments,
                 reference,
+                specializerContext,
                 callExpression
             )
             val typeParameterContext = TypeParameterContext(