+++ verik-compiler/src/main/kotlin/io/verik/compiler/specialize/TypeParameterContext.kt
@@ -22,6 +22,7 @@ import io.verik.compiler.ast.element.common.EFile
 import io.verik.compiler.ast.element.common.EReceiverExpression
 import io.verik.compiler.ast.element.common.ETypeParameter
 import io.verik.compiler.ast.element.kt.EKtClass
+import io.verik.compiler.ast.element.kt.EPrimaryConstructor
 import io.verik.compiler.ast.interfaces.TypeParameterized
 import io.verik.compiler.ast.property.Type
 import io.verik.compiler.message.Messages
@@ -54,17 +55,17 @@ data class TypeParameterContext(val typeParameterBindings: List<TypeParameterBin
         fun getFromTypeArguments(
             typeArguments: List<Type>,
             typeParameterized: TypeParameterized,
+            specializerContext: SpecializerContext,
             element: EElement
         ): TypeParameterContext {
-            typeArguments.forEach {
-                if (!it.isSpecialized())
-                    Messages.INTERNAL_ERROR.on(element, "Type argument not specialized: $it")
+            val specializedTypeArguments = typeArguments.map {
+                TypeSpecializer.specialize(it, specializerContext, element, false)
             }
             val expectedSize = typeParameterized.typeParameters.size
-            val actualSize = typeArguments.size
+            val actualSize = specializedTypeArguments.size
             return if (expectedSize == actualSize) {
                 val typeParameterBindings = typeParameterized.typeParameters
-                    .zip(typeArguments)
+                    .zip(specializedTypeArguments)
                     .map { (typeParameter, type) -> TypeParameterBinding(typeParameter, type) }
                 TypeParameterContext(typeParameterBindings)
             } else {
@@ -92,12 +93,13 @@ data class TypeParameterContext(val typeParameterBindings: List<TypeParameterBin
                     getFromTypeArguments(
                         specializedType.arguments,
                         specializedTypeReference,
+                        specializerContext,
                         receiverExpression
                     )
                 } else EMPTY
             } else {
                 val reference = receiverExpression.reference
-                if (reference is EDeclaration && reference.parent !is EFile)
+                if (reference is EDeclaration && reference !is EPrimaryConstructor && reference.parent !is EFile)
                     specializerContext.typeParameterContext
                 else EMPTY
             }