+++ verik-compiler/src/test/kotlin/io/verik/compiler/specialize/DeclarationSpecializerStageTest.kt
@@ -35,6 +35,18 @@ internal class DeclarationSpecializerStageTest : BaseTest() {
     }
 
     @Test
+    fun `specialize class type parameter cardinal function`() {
+        driveElementTest(
+            """
+                class C<N : `*`>
+                val c = C<INC<`7`>>()
+            """.trimIndent(),
+            DeclarationSpecializerStage::class,
+            "KtClass(C_N_8, C_N_8, [], [], [], 0, 0, 0, PrimaryConstructor(C_N_8, [], []), null)"
+        ) { it.findDeclaration("C_N_8") }
+    }
+
+    @Test
     fun `specialize class type parameter class`() {
         driveElementTest(
             """
@@ -48,7 +60,7 @@ internal class DeclarationSpecializerStageTest : BaseTest() {
     }
 
     @Test
-    fun `specialize class with property`() {
+    fun `specialize class with property parameterized`() {
         driveElementTest(
             """
                 class C<N : `*`> {
@@ -57,14 +69,39 @@ internal class DeclarationSpecializerStageTest : BaseTest() {
                 val c = C<`8`>()
             """.trimIndent(),
             DeclarationSpecializerStage::class,
+            "KtProperty(x, Ubit<`8`>, KtCallExpression(*), [], 0)"
+        ) { it.findDeclaration("x") }
+    }
+
+    @Test
+    fun `specialize class with call expression parameterized`() {
+        driveElementTest(
             """
-                KtClass(
-                    C_N_8, C_N_8,
-                    [KtProperty(x, Ubit<`8`>, KtCallExpression(*), [], 0)],
-                    [], [], 0, 0, 0, *, null
-                )
-            """.trimIndent()
-        ) { it.findDeclaration("C_N_8") }
+                @Suppress("MemberVisibilityCanBePrivate")
+                class C<N : `*`> {
+                    fun f(): Boolean { return false }
+                    val x = f()
+                }
+                val c = C<`8`>()
+            """.trimIndent(),
+            DeclarationSpecializerStage::class,
+            "KtProperty(x, Boolean, KtCallExpression(Boolean, f, null, [], []), [], 0)"
+        ) { it.findDeclaration("x") }
+    }
+
+    @Test
+    fun `specialize class with call expression not parameterized`() {
+        driveElementTest(
+            """
+                class D
+                class C<N : `*`> {
+                    val x = D()
+                }
+                val c = C<`8`>()
+            """.trimIndent(),
+            DeclarationSpecializerStage::class,
+            "KtProperty(x, D, KtCallExpression(D, <init>, null, [], []), [], 0)"
+        ) { it.findDeclaration("x") }
     }
 
     @Test