+++ CHANGELOG.md
@@ -8,12 +8,14 @@
 - Support tasks with return values.
 - `fork` and `join` functions.
 - Support x and z bit constants.
+- Support type alias with type parameters.
 
 ### Changed
 - Adopt `StageType` based stage sequence in importer.
 - Turn off `enableLineDirective` by default.
 - Renamed `Logical` to `Optional`.
 - Rewrote constant expression handling.
+- Merged `Optional` with `Cardinal`.
 
 ## [0.1.11]
 ### Added